requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
astroid==2.4.2
attrs==20.2.0
babel==2.9.0
brotlipy==0.7.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.7.0
colorama==0.4.4
cryptography==3.2
dnspython==2.0.0
email_validator==1.1.1
fiona==1.8.13
flask==1.1.2
flask-assets==2.0
flask-login==0.5.0
flask-paginate==0.7.0
flask-sqlalchemy==2.4.4
flask-wtf==0.14.3
flask-script==2.0.6
flask-session==0.3.2
geopandas==0.8.1
h5py==2.9.0
icc-rt==2019.0.0
pyicu==2.6
idna==2.10
intel-openmp==2020.0.133
isort==5.6.4
itsdangerous==1.1.0
jinja2==2.11.2
joblib==0.17.0
pycurl==7.43.0.6
spatialite==0.0.3
xgboost==0.90
pyxml2dict==1.0.3
lz4==2.1.10
markupsafe==1.1.1
mccabe==0.6.1
mkl==2019.0
munch==2.5.0
numpy==1.19.2
pandas==1.1.3
py-postgresql==1.2.2
pyproj==2.6.1.post1
pysocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
psycopg2-binary==2.8.6
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.2
shapely==1.7.1
six==1.15.0
sqlalchemy==1.3.20
tbb==2020.3.254
threadpoolctl==2.1.0
tiledb==0.7.0
toml==0.10.1
urllib3==1.25.11
vc==2018.7.10
webassets==2.0
werkzeug==1.0.1
win_inet_pton==1.1.0
wincertstore==0.2
wrapt==1.11.2
wtforms==2.3.3
pygeos==0.8
pygdal==2.4.2.6
psycopg2==2.8.4
flask-migrate==2.4.0
flask-script==2.0.6
python-dotenv==0.15.0
redis==3.5.3
plotly==4.13.0
swagger-ui-bundle == 0.0.8
connexion == 2.7.0
flask-swagger-ui == 3.36.0


